Finished converting css to styled-components. Converted MediaObject and TextObject into one component - BoxObject. Made components for the visual feedback blocks (On drag, on resize, on delete). Did some other stuff. My brain hurts.
