+++ components/BoxObject/index.js
@@ -1,0 +1,3 @@
+import BoxObject from "./BoxObject"
+
+export default BoxObject;
