--- components/MediaObject/index.module.css
@@ -1,5 +1,0 @@
-.mediaObject {
-    user-select: none;
-    width: 100%;
-    height: 100%;
-}
