--- components/TextObject/index.module.css
@@ -1,5 +1,0 @@
-.textObject {
-    width: 100%;
-    height: 100%;
-    user-select: none;
-}
